refactor: userSettingModal 경로 수정
