+++ src/store/use-modal-store.ts
@@ -1,7 +1,7 @@
 import { create } from 'zustand';
 import { TConfirmModalProps } from '@/components/common/modal/confirm-modal/confirm-modal';
 import { TAlertModalProps } from '@/components/common/modal/alert-modal/alert-modal';
-import { TUserSettingModalProps } from '@/components/user-setting-modal/user-setting-modal';
+import { TUserSettingModalProps } from '@/components/modal/user-setting-modal/user-setting-modal';
 
 type TModalType = 'confirm' | 'alert' | 'form' | 'user-setting';
